requirements.txt
Django==1.8.2
Fabric==1.10.2
Werkzeug==0.9.6
cuisine==0.7.8
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.2
django-extensions==1.3.8
django-grappelli==2.6.5
django-phonenumber-field==0.7.2
django-registration-redux==1.2
djangorestframework==3.1.3
ecdsa==0.13
gunicorn==19.3.0
newrelic==2.52.0.40
paramiko==1.15.2
phonenumbers==7.0.6
psycopg2==2.6.1
pycrypto==2.6.1
raven==5.3.1
six==1.9.0
sqlparse==0.1.15
unicodecsv==0.13.0
wsgiref==0.1.2
yolk==0.4.3
